requirements.txt
# FastAPI and web framework dependencies
fastapi==0.115.6
uvicorn[standard]==0.34.0
pydantic==2.10.3
pydantic[email]==2.10.3

# Authentication and security
PyJWT==2.8.0
python-multipart==0.0.12

# Core dependencies (from existing project)
PySide6==6.9.0
requests==2.32.3
python-dateutil==2.8.2
python-dotenv==1.0.1
Pillow==11.2.1
google-generativeai==0.8.3
schedule==1.2.2
requests-oauthlib==2.0.0

# Video processing dependencies
opencv-python==4.10.0.84
moviepy==1.0.3

# AI and ML dependencies
google-genai==1.19.0

# Optional dependencies
pypdf>=3.15.1
numpy<2
pytz==2023.3

# Testing dependencies
pytest==7.4.3
httpx==0.25.2
pytest-asyncio==0.21.1

# Development dependencies
black==23.11.0
isort==5.12.0
flake8==6.1.0 